Refactored build script
Refactored the build script to be more independent and easier to
configure and adapt. This meant moving paths to static variables
declared at the beginning of the file, adding a directory walker
to find .rs.in files automatically, and adding some code to write
include files pointing to the generated sources automatically.
Moreover, the script now listens to environment variables passed
by gradle to set the app package name automatically.

I have also added an assertion to check that the target is an
android triple. For cross-platform builds, this would need to
be refactored into a method under #[cfg(target_os="android")].

Finally, the gradle script has been modified to use independent
task labels and to reflect the changes to build.rs.