Fixed harlmess duplicate generated code with Vec<T> and possibly some others.
As discussed on #435, the code generation may create duplicate generated code.
Most of the time, it's harmless, but it may bring troubles later or with
other parts of code.

This happens in the following scenario when there is a function
having a Vec<T> in parameter and another one as return value:

```
    class Foo {
        fn Foo::return_is_array_u8() -> Vec<u8>;
        fn Foo::param_is_array_u8(_: Vec<u8>);
    }
```
If there is :
- only one of them,
- or a unique function having both,
the duplicate generated code is not seen.

In this fix, a `foreign_code_cache` is added to `CppContext`.
This cache is checked before writing from `cpp-include.rs` code rules.
When the code entry is already present in the cache, it's no more written
to FileWriteCache.

When collecting generated files, 2 markers are added to clearly identify
the begining and the end of each files:
* `<<< generated file: my_foo_file.h >>>`
* `>>> end of file: my_foo_file.h <<<`
Then, one can use both also to check generated file names.