jni_tests: Add test cases for String encoding.
The JNI interfaces `NewStringUTF` and `GetStringUTFChars` actually uses
"Modified UTF-8" which is similar to UTF-8 but uses a 2-byte encoding for
embedded NULL bytes and 2-byte surrogate pairs for 4-byte Unicode chars.

These are not accepted by Rust's `str`/`String` (which require full UTF-8) and
JNIs does not accept proper UTF-8 as input.

Currently passing such a string from Java to Rust panics in the unwrap in
`s.to_str().unwrap()` within `JavaString::to_str()`. We test as both `&str` and
`String` arguments.

When passing an embedded NULL from Rust to Java the string is truncated at the
first NULL.

When passing a 4-byte character from Rust to Java the result appears to be some
diffent character the returned `𐐷` becomes `ð`.

Currently all six of the cases added here fail.